+++ Character/Class/Warrior/Spells/Execute.cpp
@@ -36,18 +36,18 @@ void Execute::spell_effect() {
     // TODO: Check Execute rage loss on miss/dodge/parry
     if (result == AttackResult::MISS) {
         increment_miss();
-        warr->lose_rage(resource_cost);
+        warr->lose_rage(static_cast<unsigned>(resource_cost));
         return;
     }
     if (result == AttackResult::DODGE) {
         increment_dodge();
         warr->get_overpower_buff()->apply_buff();
-        warr->lose_rage(static_cast<int>(round(resource_cost * 0.25)));
+        warr->lose_rage(static_cast<unsigned>(round(resource_cost * 0.25)));
         return;
     }
     if (result == AttackResult::PARRY) {
         increment_parry();
-        warr->lose_rage(static_cast<int>(round(resource_cost * 0.25)));
+        warr->lose_rage(static_cast<unsigned>(round(resource_cost * 0.25)));
         return;
     }
