+++ CombatRoll/Random.cpp
@@ -1,7 +1,7 @@
 
 #include "Random.h"
 
-Random::Random(const int min_range, const int max_range):
+Random::Random(const unsigned min_range, const unsigned max_range):
     min_range(min_range),
     modulo(max_range - min_range),
     xoroshiro(new xoroshiro128plus())
@@ -11,12 +11,12 @@ Random::~Random() {
     delete xoroshiro;
 }
 
-void Random::set_gen_from_seed(const unsigned long seed) {
-    xoroshiro->set_state((uint64_t)seed);
+void Random::set_gen_from_seed(const unsigned long long seed) {
+    xoroshiro->set_state(seed);
 }
 
-int Random::get_roll(void) {
+unsigned Random::get_roll(void) {
     if (modulo == 0)
-        return min_range;
-    return (xoroshiro->next() % modulo) + min_range;
+        return static_cast<unsigned>(min_range);
+    return static_cast<unsigned>((xoroshiro->next() % modulo) + min_range);
 }
